# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Netflix Api
host: localhost:10010
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  sessionId:
    type: apiKey
    in: header
    name: "X-Session-ID"
paths:
  /signup:
    x-swagger-router-controller: user
    post:
      tags:
      - user
      description: "Creating new user"
      operationId: signupUser
      parameters:
        - name: "user"
          in: body
          description: "User object to be created"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "201":
          description: "User created"
          schema:
            $ref: "#/definitions/User"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /login:
    x-swagger-router-controller: user
    post:
      tags:
      - user
      description: "Logging in with user"
      operationId: loginUser
      parameters:
        - name: "credentials"
          in: body
          description: "Check users username and password"
          required: true
          schema:
            $ref: "#/definitions/Credentials"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Session"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /videos:
    x-swagger-router-controller: video
    get:
      tags:
      - video
      description: "Search video by title"
      operationId: searchVideo
      parameters:
        - name: title
          in: query
          required: true
          type: "string"
        - name: "X-Session-ID"
          in: header
          required: true
          type: "string"
      security:
        - sessionId: []
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/Video"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /videos/{id}/download:
    x-swagger-router-controller: video
    post:
      tags:
      - video
      description: "Download video to the users queue"
      operationId: downloadVideo
      parameters:
        - name: id
          in: path
          description: "The videos ID"
          required: true
          type: "string"
        - name: "X-Session-ID"
          in: header
          required: true
          type: "string"
      security:
        - sessionId: []
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Video"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
  /videos/queue:
    x-swagger-router-controller: video
    get:
      tags:
      - video
      description: "List all videos downloaded by the user"
      operationId: listDownloaded
      parameters:
        - name: "X-Session-ID"
          in: header
          required: true
          type: "string"
      security:
        - sessionId: []
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/Video"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
  /logout:
    x-swagger-router-controller: user
    get:
      description: "Logging out with current user"
      tags:
      - user
      operationId: logout
      parameters:
      - name: "X-Session-ID"
        in: header
        required: true
        type: "string"
      security:
        - sessionId: []
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  User:
    properties: 
      id:
        type: "number"
        description: "Number to identify user"
        example: 145
      userName:
        type: "string"
        description: "Users chosen name"
        example: "johnDoe"
      email:
        type: "string"
        description: "Users email adress"
        example: "jd23@gmail.com"
      password:
        type: "string"
        description: "Users secret password"
        example: "seeecret"
    #additionalProperties: true/false - (más property-k is lehetnek-e benne)
    required:
    - userName
    - email
    - password
  Credentials:
    required:
      - "userName"
      - "password"
    properties:
      userName:
        type: "string"
        example: "johnDoe"
      password:
        type: "string"
        example: "seeecret"
  Session:
    required:
      - sessionID
    properties:
      sessionID:
        type: "string"
        example: '2331lkgj342d'
      ExpiresAfter:
        type: "string"
        format: "date-time"
        description: "date when token expires"
  Video:
    required:
      - id
      - title
    properties:
      id:
        type: "number"
        example: "12"
      title:
        type: "string"
        example: "Matrix"
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: "string"